requirements.txt
qq-botpy
python-dotenv
sqlalchemy
apscheduler
requests
tomli